+++ packages/theme-default/modern.config.ts
@@ -17,6 +17,7 @@ const COMMON_EXTERNALS = [
   'virtual-search-hooks',
   '@rspress/runtime',
   '@theme',
+  /@theme-assets\//,
   'virtual-i18n-text',
   'virtual-prism-languages',
 ];
@@ -84,6 +85,18 @@ export default defineConfig({
       sourceMap: true,
       dts: {},
       externals: ['tailwindcss'],
+    },
+    // pre-built svg files
+    {
+      input: ['src/assets'],
+      outDir: 'dist/assets',
+      asset: {
+        svgr: true,
+        name: '[name].[ext]',
+      },
+      dts: false,
+      format: 'esm',
+      target: 'es2020',
     },
   ],
 });