+++ packages/theme-default/src/components/Search/SuggestItem.tsx
@@ -1,11 +1,11 @@
 import { isProduction } from '@rspress/runtime';
-import FileSvg from './assets/file.svg';
-import JumpSvg from './assets/jump.svg';
-import HeaderSvg from './assets/header.svg';
-import TitleSvg from './assets/title.svg';
-import { DefaultMatchResultItem, HightlightInfo } from './logic/types';
-import styles from './index.module.scss';
+import FileSvg from '@theme-assets/file';
+import JumpSvg from '@theme-assets/jump';
+import HeaderSvg from '@theme-assets/header';
+import TitleSvg from '@theme-assets/title';
 import { getSlicedStrByByteLength, removeDomain } from './logic/util';
+import styles from './index.module.scss';
+import { DefaultMatchResultItem, HightlightInfo } from './logic/types';
 
 const ICON_MAP = {
   title: TitleSvg,